Add JSON support for OpenFoam dicts.
This consists of JSON sample files in samples directory which are used
to generate a schema, that is manually modified with a title and
dropping of all required properites. The schemas as stored in the
schemas directory.

The class FoamDictJSONGenerator generates a python class with bindings
for a JSON instance for a given defined schema.