Refactor blockMesh, solver runners into a single foam cmd runner class.
The blockmesh package is deleted.
The test_mesh_runner and test_solver_runner modules are deleted.
The test_solver_runner is renamed to test_foam_cmd_runner and tests the
blockMesh command, the icoFoam command all tutorials.
The tutorial test is right now failing with 30 cases, but that should
eventually be fixed up as complete openfoam support is added.